% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootbivar.R
\name{bootbivar}
\alias{bootbivar}
\title{Bivariate bootstrap estimator of correlation, standard error, bias and confidence interval.}
\usage{
\code{bootbivar(n, x, y)}
}
\arguments{
\item{n}{The amount of bootstrap replications}

\item{x}{a vector containing observations of the first variable}

\item{y}{a vector containing observations of the second variable}
}
\value{
Correlation, estimated correlation, standard error, bias and 95% confidence interval of variables x and y.
}
\description{
Applying a simple bootstrap estimate, this function estimates the correlation of variables x and y, as well as standard error, bias and confidence interval.
}
\author{
Bjorn Helgason
}
\keyword{bias}
\keyword{bootstrap}
\keyword{confidence}
\keyword{correlation}
\keyword{error}
\keyword{interval}
\keyword{standard}

