% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.gof.r
\name{chi.gof}
\alias{chi.gof}
\title{Chi Square Goodness of Fit}
\usage{
\code{chi.gof(x,p, mehtod = "Monte Carlo")}
}
\arguments{
\item{x}{a vector containing the counts of the categorical variable.}

\item{p}{specifies the expected probability of each outcome. If not specified, assumes uniform distribution.}

\item{method}{indicates the method of computing the p-value. Use \code{"Monte Carlo"} for a Monte Carlo simulation, using \code{chi.probability}, or \code{"base"} for the R base \code{pchisq}.}
}
\value{
Returns the Chi Square Goodness of Fit value.
}
\description{
This function performs a Chi Square Goodness of Fit test, estimating the p-value, using the function \code{chi.probability}.
}
\author{
Bjorn Helgason
}
\keyword{Chi}
\keyword{Chi-squared}
\keyword{Fit}
\keyword{Goodness}
\keyword{Square}
\keyword{of}

