% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.estimator.r
\name{dense.estimator}
\alias{dense.estimator}
\title{Kernel Desnity Estimator}
\usage{
\code{dense.estimator(x, d, h, method = "naive")}
}
\arguments{
\item{x}{a vector representing a numeric variable.}

\item{d}{a vector specifying the points of evaluation. If left unspecified, evaluates at min, 1st quantile, median, mean, 3rd quantile and max of the numeric variable.}

\item{h}{specifies the bandwith used for the estimation. Defaults to Sturges' rule for Naive method and Silverman's for Kernel method.}

\item{method}{indicates whether teh naive method or the gaussian kernel is used for estimation.}
}
\value{
Returns the estimated density at points \code{d}, or at min, 1st quantile, median, mean, 3rd quantile and max of the numeric variable, if unspecified.
}
\description{
\code{dense.estimator} estimates the density of a numeric variable, evaluated at a point.
}
\author{
Bjorn Helgason
}
\keyword{density}
\keyword{estimator}
\keyword{kernel}
\keyword{naive}

