% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.plot.r
\name{dense.plot}
\alias{dense.plot}
\title{Kernel Desnity Estimate Plot}
\usage{
\code{dense.plot(x, d, h, method = "naive")}
}
\arguments{
\item{x}{a vector representing a numeric variable.}

\item{method}{indicates whether teh naive method or the gaussian kernel is used for estimation.}

\item{...}{extra parameters to be passed to \code{plot}.}

\item{d}{a vector specifying the points of evaluation. If left unspecified, evaluates at min, 1st quantile, median, mean, 3rd quantile and max of the numeric variable.}

\item{h}{specifies the bandwith used for the estimation. Defaults to Sturges' rule for Naive method and Silverman's for Kernel method.}
}
\value{
No object is returned by this function
}
\description{
Plots the estimated density of numeric variable 'x', using \code{dense.estimator}.
}
\author{
Bjorn Helgason
}
\keyword{density}
\keyword{estimator}
\keyword{kernel}
\keyword{naive}
\keyword{plot}

