% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Regression with Matrices}
\usage{
\code{linreg(formula)}
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}
}
\value{
Returns a list of residual standard error, r square, adjusted r square, f statistic, error degrees of freedom, model degrees of freedom, f p-value and the coefficent table.
}
\description{
\code{linreg} executes linear regression using matrix regression.
}
\author{
Bjorn Helgason
}
\keyword{linear}
\keyword{matrix}
\keyword{regression}

