% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcstation.R
\name{mcstation}
\alias{mcstation}
\title{Markov Chain Sationary Probability Estimator}
\usage{
\code{mcstation(p, k, n)}
}
\arguments{
\item{p}{a matrix containing transition probabilities.}

\item{k}{an integer dictating the initial state of the chain.}

\item{n}{an integer specifying the number of steps simulated}
}
\value{
Returns a table of stationary probabilities.
}
\description{
Estimates the stationary probabilities of a maarkov chain, using simulation.
}
\author{
Bjorn Helgason
}
\keyword{chain}
\keyword{estimator}
\keyword{markov}
\keyword{probability}
\keyword{stationary}

