% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throw.R
\name{throw}
\alias{throw}
\title{Throw Needles for Buffon}
\usage{
throw()
}
\arguments{
\item{l}{The length of the needles}

\item{d}{The distance between lines}

\item{n}{The amount of needles thrown}

\item{PI}{TRUE if the function should output an estimate of pi}
}
\value{
Returns the amount of needles that hit a line
}
\description{
Throw a needle and determine whether or not it lands on a line. If PI = TRUE, instead returns an estimate of pi.
}
\author{
Bjorn Helgason
}
\keyword{Buffon}
\keyword{Buffon's}
\keyword{needle}
\keyword{throw}

